Cleaning up the code and fixing a warning from pandas
