+++ dataactvalidator/validation_handlers/validationManager.py
@@ -5,13 +5,15 @@ import os
 import traceback
 import pandas as pd
 
-from pandas import isnull
 from datetime import datetime
 
 from sqlalchemy import and_, or_
 from sqlalchemy.exc import SQLAlchemyError
 
 from dataactbroker.handlers.submission_handler import populate_submission_error_info
+from dataactbroker.helpers.validation_helper import (
+    derive_fabs_awarding_sub_tier, derive_fabs_afa_generated_unique, derive_fabs_unique_award_key, derive_unique_id,
+    check_required, check_type, check_length, clean_col, clean_numbers, concat_flex)
 
 from dataactcore.aws.s3Handler import S3Handler
 from dataactcore.config import CONFIG_BROKER, CONFIG_SERVICES
@@ -22,9 +24,9 @@ from dataactcore.interfaces.function_bag import (
     populate_job_error_info, get_action_dates
 )
 
-from dataactcore.models.domainModels import matching_cars_subquery, Office, concat_display_tas_dict
+from dataactcore.models.domainModels import matching_cars_subquery, Office
 from dataactcore.models.jobModels import Submission
-from dataactcore.models.lookups import FILE_TYPE, FILE_TYPE_DICT, RULE_SEVERITY_DICT, FIELD_TYPE_DICT, FIELD_TYPE_DICT_ID
+from dataactcore.models.lookups import FILE_TYPE, FILE_TYPE_DICT, RULE_SEVERITY_DICT, FIELD_TYPE_DICT
 from dataactcore.models.validationModels import FileColumn
 from dataactcore.models.stagingModels import DetachedAwardFinancialAssistance, FlexField
 from dataactcore.models.errorModels import ErrorMetadata
@@ -41,7 +43,7 @@ from dataactvalidator.filestreaming.csvReader import CsvReader
 from dataactvalidator.filestreaming.fieldCleaner import FieldCleaner, StringCleaner
 
 from dataactvalidator.validation_handlers.errorInterface import ErrorInterface
-from dataactvalidator.validation_handlers.validator import Validator, cross_validate_sql, validate_file_by_sql
+from dataactvalidator.validation_handlers.validator import cross_validate_sql, validate_file_by_sql
 from dataactvalidator.validation_handlers.validationError import ValidationError
 
 logger = logging.getLogger(__name__)
@@ -168,7 +170,6 @@ class ValidationManager:
         job_id = job.job_id
         submission_id = job.submission_id
 
-        row_number = 1
         file_type = job.file_type.name
         validation_start = datetime.now()
 
@@ -276,20 +277,6 @@ class ValidationManager:
             # Replace whatever the user included so we're using the database headers
             data.rename(columns=reader.header_dict, inplace=True)
 
-            def clean_col(value):
-                if isnull(value) or not str(value).strip():
-                    return None
-
-                # Trim and remove extra quotes around the outside. If removing quotes and stripping leaves nothing, set
-                # it to None
-                value = str(value).strip()
-                if value.startswith('"') and value.endswith('"'):
-                    value = value[1:-1].strip()
-                    if not value:
-                        return None
-
-                return value
-
             if len(data.index) > 0:
                 data = data.applymap(clean_col)
 
@@ -306,24 +293,12 @@ class ValidationManager:
 
             flex_data = None
             if reader.flex_fields:
-                flex_data = data[list(reader.flex_fields + ['row_number'])]
-
-            # TODO: Move this
-            def concat_flex(row):
-                return ', '.join([name + ': ' + (row[name] or '') for name in sorted(row.keys()) if name is not 'row_number'])
+                flex_data = data.loc[:, list(reader.flex_fields + ['row_number'])]
 
             if flex_data is not None:
                 flex_data['concatted'] = flex_data.apply(lambda x: concat_flex(x), axis=1)
 
             data = data[list(expected_headers + ['row_number'])]
-
-            # TODO: Check if we even need this
-            def clean_numbers(value):
-                if value is not None:
-                    temp_value = value.replace(',', '')
-                    if FieldCleaner.is_numeric(temp_value):
-                        return temp_value
-                return value
 
             # Cleaning up numbers so they can be inserted properly
             for field in number_fields:
@@ -333,12 +308,10 @@ class ValidationManager:
             error_df = pd.DataFrame(columns=self.report_headers)
             warning_df = pd.DataFrame(columns=self.report_headers)
 
-            # TODO: this is temporary to make sure the file can be read the old-fashioned way until this is working
-            reader.file.seek(0)
-            reader.file.readline()
-
             # list to keep track of rows that fail validations
             error_rows = []
+
+            is_fabs = (file_type == 'fabs')
 
             # While not done, pull rows in chunks and put them into staging table if they pass the Validator
 
@@ -357,7 +330,7 @@ class ValidationManager:
             required_list = None
             type_list = None
             office_list = {}
-            if file_type == "fabs":
+            if is_fabs:
                 # create a list of all required/type labels for FABS
                 labels = sess.query(ValidationLabel).all()
                 required_list = {}
@@ -378,160 +351,19 @@ class ValidationManager:
 
             # Only do validations if it's not a D file
             if file_type not in ['award', 'award_procurement']:
-                # TODO: Move these functions
-                def derive_unique_id(row, is_fabs):
-                    if not is_fabs:
-                        return 'TAS: {}'.format(concat_display_tas_dict(row))
-
-                    return 'AssistanceTransactionUniqueKey: {}'.format(row['afa_generated_unique'])
-
-                def derive_fabs_awarding_sub_tier(row):
-                    if not row['awarding_sub_tier_agency_c']:
-                        return office_list.get(row['awarding_office_code'])
-                    return row['awarding_sub_tier_agency_c']
-
-                def derive_fabs_afa_generated_unique(row):
-                    return (row['awarding_sub_tier_agency_c'] or '-none-') + '_' + \
-                           (row['fain'] or '-none-') + '_' + \
-                           (row['uri'] or '-none-') + '_' + \
-                           (row['cfda_number'] or '-none-') + '_' + \
-                           (row['award_modification_amendme'] or '-none-')
-
-                def derive_fabs_unique_award_key(row):
-                    if str(row['record_type']) == '1':
-                        unique_award_key_list = ['ASST_AGG', row['uri'] or '-none-']
-                    else:
-                        unique_award_key_list = ['ASST_NON', row['fain'] or '-none-']
-
-                    unique_award_key_list.append(row['awarding_sub_tier_agency_c'] or '-none-')
-
-                    return '_'.join(unique_award_key_list).upper()
-
-                def apply_label(row, labels, is_fabs):
-                    if is_fabs and labels and row['Field Name'] in labels:
-                        return labels[row['Field Name']]
-                    return ''
-
-                def gather_flex_fields(row, flex_data):
-                    if flex_data is not None:
-                        return flex_data.loc[flex_data['row_number'] == row['Row Number'], 'concatted'].values[0]
-                    return ''
-
-                def valid_type(row, csv_schema):
-                    current_field = csv_schema[row['Field Name']]
-                    return Validator.check_type(row['Value Provided'], FIELD_TYPE_DICT_ID[current_field.field_types_id])
-
-                def expected_type(row, csv_schema):
-                    current_field = csv_schema[row['Field Name']]
-                    return 'This field must be a {}'.format(FIELD_TYPE_DICT_ID[current_field.field_types_id].lower())
-
-                def valid_length(row, csv_schema):
-                    current_field = csv_schema[row['Field Name']]
-                    if current_field.length:
-                        return len(row['Value Provided']) <= current_field.length
-                    return True
-
-                def expected_length(row, csv_schema):
-                    current_field = csv_schema[row['Field Name']]
-                    return 'Max length: {}'.format(current_field.length)
-
-                def update_field_name(row, short_cols):
-                    if row['Field Name'] in short_cols:
-                        return short_cols[row['Field Name']]
-                    return row['Field Name']
-
-                def add_field_name_to_value(row):
-                    return row['Field Name'] + ': ' + row['Value Provided']
-
-                def check_required(data, required, required_labels, report_headers, short_cols, flex_data, is_fabs):
-                    req_data = data[required + ['row_number', 'unique_id']]
-                    errors = pd.melt(req_data, id_vars=['row_number', 'unique_id'], value_vars=required,
-                                     var_name='Field Name', value_name='Value Provided')
-                    errors = errors[errors['Value Provided'].isnull()]
-                    errors.rename(columns={'row_number': 'Row Number', 'unique_id': 'Unique ID'}, inplace=True)
-                    errors = errors.reset_index()
-                    errors['Value Provided'] = ''
-                    errors['Error Message'] = ValidationError.requiredErrorMsg
-                    errors['Expected Value'] = '(not blank)'
-                    errors['Difference'] = ''
-                    if not errors.empty:
-                        errors['Rule Label'] = errors.apply(lambda x: apply_label(x, required_labels, is_fabs), axis=1)
-                        errors['Flex Field'] = errors.apply(lambda x: gather_flex_fields(x, flex_data), axis=1)
-                        errors['Field Name'] = errors.apply(lambda x: update_field_name(x, short_cols), axis=1)
-                    else:
-                        errors['Rule Label'] = ''
-                        errors['Flex Field'] = ''
-                    errors = errors[report_headers]
-                    errors['error_type'] = ValidationError.requiredError
-                    return errors
-
-                def check_type(data, type_fields, type_labels, report_headers, csv_schema, short_cols, flex_data,
-                               is_fabs):
-                    type_data = data[type_fields + ['row_number', 'unique_id']]
-                    errors = pd.melt(type_data, id_vars=['row_number', 'unique_id'], value_vars=type_fields,
-                                     var_name='Field Name', value_name='Value Provided')
-                    errors = errors[~errors['Value Provided'].isnull()]
-                    if not errors.empty:
-                        errors['matches_type'] = errors.apply(lambda x: valid_type(x, csv_schema), axis=1)
-                        errors = errors[~errors['matches_type']]
-                        errors.drop(['matches_type'], axis=1, inplace=True)
-                    errors.rename(columns={'row_number': 'Row Number', 'unique_id': 'Unique ID'}, inplace=True)
-                    errors = errors.reset_index()
-                    errors['Error Message'] = ValidationError.typeErrorMsg
-                    errors['Difference'] = ''
-                    if not errors.empty:
-                        errors['Expected Value'] = errors.apply(lambda x: expected_type(x, csv_schema), axis=1)
-                        errors['Rule Label'] = errors.apply(lambda x: apply_label(x, type_labels, is_fabs), axis=1)
-                        errors['Flex Field'] = errors.apply(lambda x: gather_flex_fields(x, flex_data), axis=1)
-                        errors['Field Name'] = errors.apply(lambda x: update_field_name(x, short_cols), axis=1)
-                        errors['Value Provided'] = errors.apply(lambda x: add_field_name_to_value(x), axis=1)
-                    else:
-                        errors['Expected Value'] = ''
-                        errors['Rule Label'] = ''
-                        errors['Flex Field'] = ''
-                    errors = errors[report_headers]
-                    errors['error_type'] = ValidationError.typeError
-                    return errors
-
-                def check_length(data, length_fields, report_headers, csv_schema, short_cols, flex_data,
-                                 type_error_rows):
-                    length_data = data[length_fields + ['row_number', 'unique_id']]
-                    errors = pd.melt(length_data, id_vars=['row_number', 'unique_id'], value_vars=length_fields,
-                                     var_name='Field Name', value_name='Value Provided')
-                    errors = errors[~errors['Value Provided'].isnull() & ~errors['row_number'].isin(type_error_rows)]
-                    if not errors.empty:
-                        errors['valid_length'] = errors.apply(lambda x: valid_length(x, csv_schema), axis=1)
-                        errors = errors[~errors['valid_length']]
-                        errors.drop(['valid_length'], axis=1, inplace=True)
-                    errors.rename(columns={'row_number': 'Row Number', 'unique_id': 'Unique ID'}, inplace=True)
-                    errors = errors.reset_index()
-                    errors['Error Message'] = ValidationError.lengthErrorMsg
-                    errors['Difference'] = ''
-                    errors['Rule Label'] = ''
-                    if not errors.empty:
-                        errors['Expected Value'] = errors.apply(lambda x: expected_length(x, csv_schema), axis=1)
-                        errors['Flex Field'] = errors.apply(lambda x: gather_flex_fields(x, flex_data), axis=1)
-                        errors['Field Name'] = errors.apply(lambda x: update_field_name(x, short_cols), axis=1)
-                        errors['Value Provided'] = errors.apply(lambda x: add_field_name_to_value(x), axis=1)
-                    else:
-                        errors['Expected Value'] = ''
-                        errors['Flex Field'] = ''
-                    errors = errors[report_headers]
-                    errors['error_type'] = ValidationError.lengthError
-                    return errors
-
-                if file_type == 'fabs':
+                if is_fabs:
                     data['is_valid'] = True
-                    data['awarding_sub_tier_agency_c'] = data.apply(lambda x: derive_fabs_awarding_sub_tier(x), axis=1)
+                    data['awarding_sub_tier_agency_c'] = data.apply(lambda x:
+                                                                    derive_fabs_awarding_sub_tier(x, office_list),
+                                                                    axis=1)
                     data['afa_generated_unique'] = data.apply(lambda x: derive_fabs_afa_generated_unique(x), axis=1)
                     data['unique_award_key'] = data.apply(lambda x: derive_fabs_unique_award_key(x), axis=1)
 
-                data['unique_id'] = data.apply(lambda x: derive_unique_id(x, file_type == 'fabs'), axis=1)
+                data['unique_id'] = data.apply(lambda x: derive_unique_id(x, is_fabs), axis=1)
 
                 # Separate each of the checks to their own dataframes, then concat them together
                 req_errors = check_required(data, required_fields, required_list, self.report_headers,
-                                            self.short_to_long_dict[job.file_type_id], flex_data,
-                                            is_fabs=(file_type == 'fabs'))
+                                            self.short_to_long_dict[job.file_type_id], flex_data, is_fabs=is_fabs)
                 logger.info({
                     'message': 'Finished required checks',
                     'message_type': 'ValidatorInfo',
@@ -544,7 +376,7 @@ class ValidationManager:
                 })
                 type_errors = check_type(data, number_fields + boolean_fields, type_list, self.report_headers,
                                          csv_schema, self.short_to_long_dict[job.file_type_id], flex_data,
-                                         is_fabs=(file_type == 'fabs'))
+                                         is_fabs=is_fabs)
                 logger.info({
                     'message': 'Finished type checks',
                     'message_type': 'ValidatorInfo',
@@ -569,7 +401,7 @@ class ValidationManager:
                     'start_time': loading_start
                 })
 
-                if file_type == 'fabs':
+                if is_fabs:
                     error_dfs = [req_errors, type_errors, length_errors]
                     warning_dfs = [pd.DataFrame(columns=self.report_headers)]
                 else:
@@ -594,7 +426,8 @@ class ValidationManager:
                 total_warnings.drop(['error_type'], axis=1, inplace=True, errors='ignore')
 
                 total_errors.to_csv(error_file_path, columns=self.report_headers, index=False, quoting=csv.QUOTE_ALL)
-                total_warnings.to_csv(warning_file_path, columns=self.report_headers, index=False, quoting=csv.QUOTE_ALL)
+                total_warnings.to_csv(warning_file_path, columns=self.report_headers, index=False,
+                                      quoting=csv.QUOTE_ALL)
 
                 # Remove type error rows from original dataframe
                 data = data[~data['row_number'].isin(type_error_rows)]
@@ -679,7 +512,7 @@ class ValidationManager:
 
             # Update fabs is_valid rows where applicable
             # Update submission to include action dates where applicable
-            if file_type == "fabs":
+            if is_fabs:
                 sess.query(DetachedAwardFinancialAssistance).\
                     filter(DetachedAwardFinancialAssistance.row_number.in_(error_rows_unique),
                            DetachedAwardFinancialAssistance.submission_id == submission_id).\
@@ -703,7 +536,7 @@ class ValidationManager:
             # Update error info for submission
             populate_job_error_info(job)
 
-            if file_type == "fabs":
+            if is_fabs:
                 # set number of errors and warnings for detached submission
                 populate_submission_error_info(submission_id)
